requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.0.2
gpiozero==1.4.1
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pkg-resources==0.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
RPi.GPIO==0.6.5
six==1.12.0
Werkzeug==0.14.1
